+++ R/zzz.R
@@ -2,14 +2,6 @@ utils::globalVariables(".data")
 
 
 .onLoad <- function(libname, pkgname) {
-  .objects <<- memoise::memoize(.objects, cache = cachem::cache_mem(max_age = 3600))
-  oisstfiles <<- memoise::memoize(oisstfiles, cache = cachem::cache_mem(max_age = 3600))
-  .curated_objects <<- memoise::memoize(.curated_objects, cache = cachem::cache_mem(max_age = 3600))
-  nsidc25kmSfiles <<- memoise::memoize(nsidc25kmSfiles, cache = cachem::cache_mem(max_age = 3600 * 12))
-  nsidc25kmNfiles <<- memoise::memoize(nsidc25kmNfiles, cache = cachem::cache_mem(max_age = 3600 * 12))
-  readoisst <<- memoise::memoize(readoisst, cache = cachem::cache_mem(max_age = 3600 * 24))
-  readnsidc25kmN <<- memoise::memoize(readnsidc25kmN, cache = cachem::cache_mem(max_age = 3600 * 24))
-sooty_files <<- memoise::memoize(sooty_files, cache = cachem::cache_mem(max_age = 3600))
 
   Sys.setenv("AWS_NO_SIGN_REQUEST" = "YES")
   Sys.setenv("AWS_S3_ENDPOINT" = "projects.pawsey.org.au")